requirements.txt
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Django==5.1.1
djangorestframework==3.15.2
exceptiongroup==1.2.2
grpcio==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jiter==0.5.0
joblib==1.4.2
nltk==3.9.1
numpy==2.1.1
openai==1.46.1
outcome==1.3.0.post0
packaging==24.1
portalocker==2.10.1
protobuf==5.28.2
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dotenv==1.0.1
qdrant-client==1.11.2
regex==2024.9.11
requests==2.32.3
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
